refactor: extract resume task helpers to lib/resume_task_helpers.ts
Extract 9 functions and 1 type from resume_tasks.ts into a dedicated
helper module following the established extraction pattern. Reduces
resume_tasks.ts by ~151 lines with backward-compatible re-exports.
